# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ExamenTAP API",
        "description": "Documentación de la API REST del backend ExamenTAP.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "API server"
        }
    ],
    "paths": {
...
        "/usuarios": {
            "get": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Lista usuarios con paginación, búsqueda y ordenamiento",
                "description": "Devuelve un listado paginado de usuarios. Permite buscar por código, nombre o correo electrónico y ordenar por cualquiera de las columnas visibles. Requiere el permiso \"consultar\" en la sección \"usuarios\".",
                "operationId": "usuariosIndex",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort_field",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "default": "created_at",
                            "enum": [
                                "id",
                                "name",
                                "email",
                                "created_at"
                            ]
                        }
                    },
                    {
                        "name": "sort_direction",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "default": "desc",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listado de usuarios",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UsuarioListResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Sin permiso para consultar usuarios",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Crea un nuevo usuario",
                "description": "Crea un usuario con su fotografía de perfil (almacenada en Firebase Storage), valida que el correo sea único y registra el alta en la bitácora de auditoría. Si el sistema no tiene ningún usuario registrado, este endpoint es público y el usuario creado se registra automáticamente con el perfil Administrador (modo de inicialización del sistema). Una vez que existe al menos un usuario, requiere autenticación y el permiso \"crear\" en la sección \"usuarios\".",
                "operationId": "usuariosStore",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreUsuarioRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuario creado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UsuarioResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado (el sistema ya tiene usuarios registrados)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Sin permisos para crear usuarios",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiValidationErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/usuarios/{usuario}": {
            "get": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Obtiene el detalle de un usuario",
                "description": "Devuelve la información completa de un usuario, incluyendo los perfiles asignados y las secciones a las que tiene acceso (agregadas de todos sus perfiles). Requiere el permiso \"consultar\" en la sección \"usuarios\".",
                "operationId": "usuariosShow",
                "parameters": [
                    {
                        "name": "usuario",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalle de usuario",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UsuarioDetailResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Sin permiso para consultar usuarios",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Usuario no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Actualiza un usuario existente",
                "description": "Actualiza los datos de un usuario. Captura la información anterior antes de aplicar los cambios y registra la edición en la bitácora de auditoría con el detalle antes/después. Al enviarse como multipart/form-data, debe incluir el campo _method=PUT. Requiere el permiso \"editar\" en la sección \"usuarios\".",
                "operationId": "usuariosUpdate",
                "parameters": [
                    {
                        "name": "usuario",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateUsuarioRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuario actualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UsuarioResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Sin permiso para editar usuarios",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Usuario no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiValidationErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Elimina un usuario",
                "description": "Elimina físicamente el usuario de Firebase Firestore. Requiere el permiso \"eliminar\" en la sección \"usuarios\".",
                "operationId": "usuariosDestroy",
                "parameters": [
                    {
                        "name": "usuario",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuario eliminado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiSuccessResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Sin permiso para eliminar usuarios",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Usuario no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/usuarios/export/pdf": {
            "get": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Exporta el listado de usuarios a PDF",
                "description": "Genera un PDF con título, fecha de generación y una tabla con Código, Usuario, Nombre y Fecha de creación. Requiere el permiso \"consultar\" en la sección \"usuarios\".",
                "operationId": "usuariosExportPdf",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Archivo PDF",
                        "content": {
                            "application/pdf": {}
                        }
                    },
                    "403": {
                        "description": "Sin permiso para consultar usuarios",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/usuarios/export/excel": {
            "get": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Exporta el listado de usuarios a Excel",
                "description": "Genera un archivo XLSX con las columnas Código, Usuario, Nombre y Fecha de creación. Requiere el permiso \"consultar\" en la sección \"usuarios\".",
                "operationId": "usuariosExportExcel",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Archivo Excel",
                        "content": {
                            "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet": {}
                        }
                    },
                    "403": {
                        "description": "Sin permiso para consultar usuarios",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "ApiErrorResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "status": {
                        "type": "string",
                        "example": "error"
                    },
                    "message": {
                        "type": "string",
                        "example": "Las credenciales proporcionadas son incorrectas."
                    },
                    "data": {
                        "type": "object",
                        "example": null,
                        "nullable": true
                    },
                    "errors": {
                        "type": "object",
                        "example": null,
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "Pagination": {
                "properties": {
                    "page": {
                        "type": "integer",
                        "example": 1
                    },
                    "per_page": {
                        "type": "integer",
                        "example": 10
                    },
                    "total": {
                        "type": "integer",
                        "example": 42
                    },
                    "total_pages": {
                        "type": "integer",
                        "example": 5
                    }
                },
                "type": "object"
            },
            "UsuarioListResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "status": {
                        "type": "string",
                        "example": "success"
                    },
                    "message": {
                        "type": "string",
                        "example": "Usuarios obtenidos correctamente."
                    },
                    "data": {
                        "properties": {
                            "items": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/Usuario"
                                }
                            },
                            "pagination": {
                                "$ref": "#/components/schemas/Pagination"
                            }
                        },
                        "type": "object"
                    },
                    "errors": {
                        "type": "object",
                        "example": null,
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "UsuarioResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "status": {
                        "type": "string",
                        "example": "success"
                    },
                    "message": {
                        "type": "string",
                        "example": "Usuario creado correctamente."
                    },
                    "data": {
                        "$ref": "#/components/schemas/Usuario"
                    },
                    "errors": {
                        "type": "object",
                        "example": null,
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "StoreUsuarioRequest": {
                "required": [
                    "name",
                    "email",
                    "password",
                    "foto"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Juan Pérez"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "juan.perez@example.com"
                    },
                    "password": {
                        "type": "string",
                        "format": "password",
                        "example": "Secret123!"
                    },
                    "telefono": {
                        "type": "string",
                        "example": "+521234567890",
                        "nullable": true
                    },
                    "foto": {
                        "type": "string",
                        "format": "binary"
                    },
                    "perfiles": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}